requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
behave==1.2.6
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssselect==1.2.0
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
exceptiongroup==1.2.2
execnet==2.1.1
fasteners==0.19
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud==0.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflect==7.3.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mysql-connector-python==9.0.0
num2words==0.5.13
numpy==2.1.2
openai==1.43.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parse==1.20.2
parse_type==0.6.3
pdbp==1.5.4
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.4
pikepdf==9.2.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.0
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
Pygments==2.18.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pynose==1.5.2
pyotp==2.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.3.2
pytest-html==2.0.1
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.31.0
rich==13.8.0
rsa==4.9
sbvirtualdisplay==1.3.0
selenium==4.24.0
seleniumbase==4.30.3
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
tabcompleter==1.3.3
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.4
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
wkhtmltopdf==0.2
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
